// src/components/small-hero/small-hero.interface.ts
export interface SmallHeroProps {
    headline: string;
    subline?: string;
    description?: string;
    imageSrc?: string;
    imageAlt?: string;
    primaryButtonLabel?: string;
    secondaryButtonLabel?: string;
    onClickPrimaryButton?: () => void;
    onClickSecondaryButton?: () => void;
    backgroundVariant?: 'indigo' | 'gradient';
}